requirements.txt
wavelink
discord
colorama
mysql-connector-python
jishaku
Pillow
asyncio
requests
requests_oauthlib
flask